refactor MoveCharacteristic

+++ src/model/field/fieldObjects/robot/moveCharacteristics/SlipperinessCharacteristic.java
@@ -1,7 +1,7 @@
 package model.field.fieldObjects.robot.moveCharacteristics;
 
 public class SlipperinessCharacteristic extends MoveCharacteristic {
-    public SlipperinessCharacteristic(double coefficient) {
+    public SlipperinessCharacteristic(int coefficient) {
         super(coefficient);
     }
 }
