+++ src/model/field/fieldObjects/robot/moveCharacteristics/ViscosityCharacteristic.java
@@ -1,7 +1,7 @@
 package model.field.fieldObjects.robot.moveCharacteristics;
 
 public class ViscosityCharacteristic extends MoveCharacteristic {
-    public ViscosityCharacteristic(double coefficient) {
+    public ViscosityCharacteristic(int coefficient) {
         super(coefficient);
     }
 }
